First pass at parser. Set foundation; Ready to Extend Functionality.
This is likely the beginning of a project with a fair bit of growth-potential.
The class inside of pa_parser.py should be regarded as an idea more than a
finished product, naturally. Some of it's methods have generic elements that
will extend to other features as the need arises. I will attempt to avoid
premature optimization, instead modifying the code, adding and refactoring
classes and building out tests when I have a better idea of how the potential
package would be structured, which will hopefully be soon.

All that being said, a little foresight goes a long way. Going forward, features
should be built to address a small set of problems rather than a single file
stream. This approach will hopefully keep things DRY!