Updated Parser for new Constraints.
The parser now disregards any data that is associated with a date more than
forty-five days into the future. This is to account for a bug in MAS that
apparently pulls fake future data for some reason. Likewise, the parser is also
unfriendly towards NaN values. If it encounters a blank cell in the pick_date
column of a PA file, it goes on its way and is happier for it.

One thing worth looking out for: taking PA_05.15.29.xlsx as an example, some of
the values in A13-A are not in chronological order. The parser is not currently
built to track chronology. Because it appears that pick_dates are tracked in a
running-total fashion, I imagine this might be a problem.

Although the adjacent ship-date column appears to be in proper chronological
order, I am confused by the alternating nature of pick and ship dates. On some
occasions, it seems that items are shipped before they are picked, on other
occasions, vice versa. This doesn't make immediate sense to me.