A Fistful of Excel
Good opportunity to learn more about pandas dataframes. I still have a fair bit
of learnin' on that front though! Main takeaways: the apply function is
fantastic, but there are others that do the trick also. The 'scraper' in its
current iteration is inelegant, with many methods that simply need to be
deleted, and others that should be taken private. I'm working towards a more
elegant approach, but I'm not there yet...