Added in Flask.
Flask is awesome. You can just turn a project into an api! It will communicate
with the React App for Alamance Foods. Currently, it only knows how to respond
to a single file being uploaded by storing the file and sending a json object
that contains the file name to be used as an identifier in a consequent GET
request.

TL;DR- It catches the ball and passes it back.