Not to shabby...
I built the package out and pushed the wheel on up to the server only to realize
that I was cutting a bit too many corners to pass muster. The project is much
better suited for expansion now that I've got everything interfacing from behind
a centralized flask app. I look forward to expanding on it. Thankfully, it
should be easy to push changes, given that I'm working with wheels instead of
entire repos!