Fixed Type Mismatch In Flask parser.py delta_list sorting logic.
I was sorting the list of deltas in reverse order to ensure that the furthest
date range got sorted first. Because the input values were in fact
strings instead of ints, the list was not sorted in descending numeric order, as
was the intention of the logic. The problem has been corrected by parsing the
strings into int types.